% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elematrix.R
\name{elematrix}
\alias{elematrix}
\title{Elementary matrix generation}
\usage{
elematrix(n = stop("n must be specified"), i = stop("i must be specified"),
  j = NULL, alpha = NULL)
}
\arguments{
\item{n}{dimnension of the square matrix}

\item{i}{row number}

\item{j}{column number}

\item{alpha}{a scalar}
}
\description{
Elementary matrix generation (Salce pag 46)
}
