% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{equal_matrix}
\alias{equal_matrix}
\title{check (matrix) equality in a safe manner}
\usage{
equal_matrix(x, y)
}
\arguments{
\item{x}{the first matrix}

\item{y}{the second matrix}

\item{...}{arguments to be tested}
}
\description{
check equality in safe manner
}
