% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{i}
\alias{i}
\title{Immaginary unit}
\format{An object of class \code{complex} of length 1.}
\usage{
i
}
\description{
Immaginary unit made to simplify writing complex numbers
}
\keyword{datasets}
