% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_square}
\alias{is_square}
\title{Check wheter a matrix like object is square}
\usage{
is_square(x)
}
\arguments{
\item{x}{a matrix-like object}
}
\description{
Check wheter a matrix like object (has ncol and nrow) is square
}
