% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{same_dimensions}
\alias{same_dimensions}
\title{check if tho matrix have same dimensions (nrow, ncol)}
\usage{
same_dimensions(x, y)
}
\arguments{
\item{x}{the first matrix}

\item{y}{the second matrix}
}
\description{
check if tho matrix have same dimensions (nrow, ncol)
}
